move our tests into color.test.ts
